+++ src/main/java/com/codetaylor/mc/pyrotech/modules/core/block/BlockGloamberryBush.java
@@ -1,37 +1,34 @@
 package com.codetaylor.mc.pyrotech.modules.core.block;
 
 import com.codetaylor.mc.athenaeum.util.RandomHelper;
-import com.codetaylor.mc.athenaeum.util.StackHelper;
-import com.codetaylor.mc.pyrotech.library.spi.block.IBlockShearable;
 import com.codetaylor.mc.pyrotech.modules.core.ModuleCore;
 import com.codetaylor.mc.pyrotech.modules.core.ModuleCoreConfig;
 import com.codetaylor.mc.pyrotech.modules.core.block.spi.BlockBushBase;
 import com.codetaylor.mc.pyrotech.modules.core.network.SCPacketParticleCombust;
+import com.codetaylor.mc.pyrotech.modules.core.network.SCPacketParticleGloamberry;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.Items;
 import net.minecraft.init.SoundEvents;
+import net.minecraft.item.EnumDyeColor;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
-import net.minecraft.util.EnumParticleTypes;
 import net.minecraft.util.SoundCategory;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
-import net.minecraft.world.WorldServer;
 import net.minecraftforge.common.ForgeHooks;
 
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.Random;
 
-public class BlockPyroberryBush
-    extends BlockBushBase
-    implements IBlockShearable {
+public class BlockGloamberryBush
+    extends BlockBushBase {
 
-  public static final String NAME = "pyroberry_bush";
+  public static final String NAME = "gloamberry_bush";
 
   // ---------------------------------------------------------------------------
   // - Implementation
@@ -40,13 +37,13 @@ public class BlockPyroberryBush
   @Override
   public boolean isValidBlock(IBlockState blockState) {
 
-    return (blockState.getMaterial() == Material.SAND);
+    return (blockState.getMaterial() == Material.GROUND);
   }
 
   @Override
   public Item getSeedItem() {
 
-    return ModuleCore.Items.PYROBERRY_SEEDS;
+    return ModuleCore.Items.GLOAMBERRY_SEEDS;
   }
 
   // ---------------------------------------------------------------------------
@@ -72,8 +69,8 @@ public class BlockPyroberryBush
         case 6:
         case 7:
         default:
-          this.spawnFireParticles(pos, world.provider.getDimension());
-          world.playSound(null, pos.getX(), pos.getY(), pos.getZ(), SoundEvents.BLOCK_FIRE_AMBIENT, SoundCategory.BLOCKS, 1, 1);
+          this.spawnParticles(pos, world.provider.getDimension());
+          this.playSound(world, pos);
       }
 
     }
@@ -91,7 +88,8 @@ public class BlockPyroberryBush
       ItemStack itemStack = player.getHeldItemMainhand();
       Item item = itemStack.getItem();
 
-      if (item == Items.BLAZE_POWDER) {
+      if (item == Items.DYE
+          && EnumDyeColor.CYAN == EnumDyeColor.byDyeDamage(itemStack.getMetadata())) {
         int age = this.getAge(blockState);
 
         if (age < this.getMaxAge()
@@ -102,19 +100,19 @@ public class BlockPyroberryBush
           }
 
           world.setBlockState(pos, this.withAge(age + 1), 2);
-          world.playSound(null, pos.getX(), pos.getY(), pos.getZ(), SoundEvents.BLOCK_FIRE_AMBIENT, SoundCategory.BLOCKS, 1, 1);
-          this.spawnFireParticles(pos, world.provider.getDimension());
+          this.playSound(world, pos);
+          this.spawnParticles(pos, world.provider.getDimension());
           ForgeHooks.onCropsGrowPost(world, pos, blockState, world.getBlockState(pos));
 
         } else if (this.isMaxAge(blockState)) {
-          this.spawnFireParticles(pos, world.provider.getDimension());
+          this.spawnParticles(pos, world.provider.getDimension());
           world.createExplosion(null, pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5, 1.5f, true);
         }
 
         return true;
 
       } else if (item == Items.AIR && this.isMaxAge(blockState)) {
-        this.spawnFireParticles(pos, world.provider.getDimension());
+        this.spawnParticles(pos, world.provider.getDimension());
         world.createExplosion(null, pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5, 1.5f, true);
         return true;
       }
@@ -130,7 +128,7 @@ public class BlockPyroberryBush
     super.onBlockClicked(world, pos, player);
 
     if (this.getAge(world.getBlockState(pos)) == this.getMaxAge()) {
-      this.spawnFireParticles(pos, world.provider.getDimension());
+      this.spawnParticles(pos, world.provider.getDimension());
     }
   }
 
@@ -142,26 +140,18 @@ public class BlockPyroberryBush
 
     if (RandomHelper.random().nextFloat() < 0.05
         && this.getAge(world.getBlockState(pos)) == this.getMaxAge()) {
-      this.spawnFireParticles(pos, world.provider.getDimension());
+      this.spawnParticles(pos, world.provider.getDimension());
     }
   }
 
-  @Override
-  public void onSheared(World world, BlockPos pos, IBlockState blockState, ItemStack itemStack, EntityPlayer player) {
+  private void playSound(World world, BlockPos pos) {
 
-    if (this.isMaxAge(blockState)) {
-      world.playSound(null, pos.getX(), pos.getY(), pos.getZ(), SoundEvents.ENTITY_SHEEP_SHEAR, SoundCategory.BLOCKS, 1, 1);
-      world.setBlockState(pos, this.withAge(this.getMaxAge() - 1), 3);
-      itemStack.damageItem(1, player);
-      this.spawnFireParticles(pos, player.dimension);
-      world.playSound(null, pos.getX(), pos.getY(), pos.getZ(), SoundEvents.BLOCK_FIRE_AMBIENT, SoundCategory.BLOCKS, 1, 1);
-      StackHelper.spawnStackOnTop(world, new ItemStack(ModuleCore.Items.PYROBERRIES, RandomHelper.random().nextInt(3) + 1), pos, 0.5);
-    }
+    world.playSound(null, pos.getX(), pos.getY(), pos.getZ(), SoundEvents.BLOCK_NOTE_PLING, SoundCategory.BLOCKS, 1, 1);
   }
 
-  private void spawnFireParticles(BlockPos pos, int dimension) {
+  private void spawnParticles(BlockPos pos, int dimension) {
 
-    ModuleCore.PACKET_SERVICE.sendToAllAround(new SCPacketParticleCombust(pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5, 0.5, 0.5, 0.5), dimension, pos);
+    ModuleCore.PACKET_SERVICE.sendToAllAround(new SCPacketParticleGloamberry(pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5, 0.5, 0.5, 0.5), dimension, pos);
   }
 
   // ---------------------------------------------------------------------------
@@ -176,38 +166,13 @@ public class BlockPyroberryBush
 
     if (world.canSeeSky(pos)) {
 
-      if (world.isRainingAt(pos.up())) {
-        // Reduce age if raining.
-        int age = this.getAge(blockState);
-
-        if (age > 4
-            && ForgeHooks.onCropsGrowPre(world, pos, blockState, rand.nextFloat() < ModuleCoreConfig.PYROBERRY_BUSH.RAIN_GROWTH_REVERT_CHANCE)) {
-          world.setBlockState(pos, this.withAge(age - 1), 2);
-          ForgeHooks.onCropsGrowPost(world, pos, blockState, world.getBlockState(pos));
-        }
-
-        if (!world.isRemote) {
-
-          for (int i = 0; i < 8; i++) {
-            double offsetX = (RandomHelper.random().nextDouble() * 2.0 - 1.0) * 0.4;
-            double offsetY = (RandomHelper.random().nextDouble() * 2.0 - 1.0) * 0.4;
-            double offsetZ = (RandomHelper.random().nextDouble() * 2.0 - 1.0) * 0.4;
-            double x = pos.getX() + 0.5 + offsetX;
-            double y = pos.getY() + 0.6 + offsetY;
-            double z = pos.getZ() + 0.5 + offsetZ;
-            ((WorldServer) world).spawnParticle(RandomHelper.random().nextFloat() < 0.25 ? EnumParticleTypes.SMOKE_LARGE : EnumParticleTypes.SMOKE_NORMAL, x, y, z, 4, 0.0, 0.0, 0.0, 0.0);
-          }
-
-          world.playSound(null, pos.getX(), pos.getY(), pos.getZ(), SoundEvents.BLOCK_FIRE_EXTINGUISH, SoundCategory.BLOCKS, 0.5f, 1);
-        }
-
-      } else if (world.isDaytime()) {
-        // Increase age if is daytime.
+      if (!world.isDaytime()) {
+        // Increase age if is nighttime.
         int age = this.getAge(blockState);
         boolean grew = false;
 
         if (age < this.getMaxAge()) {
-          double chance = (age == this.getMaxAge() - 1) ? ModuleCoreConfig.PYROBERRY_BUSH.BERRY_GROWTH_CHANCE : ModuleCoreConfig.PYROBERRY_BUSH.GROWTH_CHANCE;
+          double chance = (age == this.getMaxAge() - 1) ? ModuleCoreConfig.GLOAMBERRY_BUSH.BERRY_GROWTH_CHANCE : ModuleCoreConfig.GLOAMBERRY_BUSH.GROWTH_CHANCE;
 
           if (ForgeHooks.onCropsGrowPre(world, pos, blockState, rand.nextFloat() < chance)) {
             world.setBlockState(pos, this.withAge(age + 1), 2);
@@ -217,8 +182,8 @@ public class BlockPyroberryBush
         }
 
         if ((age > 2 && grew) || (this.isMaxAge(blockState) && rand.nextFloat() < 0.5)) {
-          this.spawnFireParticles(pos, world.provider.getDimension());
-          world.playSound(null, pos.getX(), pos.getY(), pos.getZ(), SoundEvents.BLOCK_FIRE_AMBIENT, SoundCategory.BLOCKS, 1, 1);
+          this.spawnParticles(pos, world.provider.getDimension());
+          this.playSound(world, pos);
         }
       }
 
@@ -228,7 +193,7 @@ public class BlockPyroberryBush
 
       if (age > 4) {
 
-        if (ForgeHooks.onCropsGrowPre(world, pos, blockState, rand.nextFloat() < ModuleCoreConfig.PYROBERRY_BUSH.OBSTRUCTED_GROWTH_REVERT_CHANCE)) {
+        if (ForgeHooks.onCropsGrowPre(world, pos, blockState, rand.nextFloat() < ModuleCoreConfig.GLOAMBERRY_BUSH.OBSTRUCTED_GROWTH_REVERT_CHANCE)) {
           world.setBlockState(pos, this.withAge(age - 1), 2);
           ForgeHooks.onCropsGrowPost(world, pos, blockState, world.getBlockState(pos));
         }