+++ src/main/java/com/codetaylor/mc/pyrotech/modules/core/init/ItemInitializer.java
@@ -40,6 +40,8 @@ public final class ItemInitializer {
     registry.registerItem(new ItemPyroberrySeeds(), ItemPyroberrySeeds.NAME);
     registry.registerItem(new ItemPyroberries(), ItemPyroberries.NAME);
     registry.registerItem(new ItemPyroberryCocktail(), ItemPyroberryCocktail.NAME);
+
+    registry.registerItem(new ItemGloamberrySeeds(), ItemGloamberrySeeds.NAME);
   }
 
   @SideOnly(Side.CLIENT)
@@ -73,7 +75,9 @@ public final class ItemInitializer {
           ModuleCore.Items.PYROBERRY_SEEDS,
           ModuleCore.Items.PYROBERRIES,
           ModuleCore.Items.PYROBERRY_WINE,
-          ModuleCore.Items.PYROBERRY_COCKTAIL
+          ModuleCore.Items.PYROBERRY_COCKTAIL,
+
+          ModuleCore.Items.GLOAMBERRY_SEEDS
       );
 
       // Material