+++ docs/changelog.md
@@ -73,8 +73,13 @@
     * Pyroberry Bush
         * Spawns in desert
         * Drops Pyroberries when sheared at max growth stage
+    * Pyroberry Seeds
+    * Pyroberries
     * Pyroberry Wine
     * Pyroberry Cocktail
+    * Gloamberry Bush
+        * Drops Gloamberries when ripe
+    * Gloamberry Seeds
     * Ore Compat:
         * Bloom yield and slag bloom yield can now be adjusted in the ore compat json files [(#310)](https://github.com/codetaylor/pyrotech-1.12/issues/310)
     * Config option to:
@@ -298,6 +303,8 @@
         * `tile.fluid.pyroberry_wine.name`
         * `item.pyrotech.pyroberry.wine.name`
         * `item.pyrotech.pyroberry.cocktail.name`
+        * `tile.pyrotech.gloamberry_bush.name`
+        * `item.pyrotech.gloamberry.seeds.name`
 
     * **Changed Lang Keys:**
         * `tile.pyrotech.stone_bricks.name` -> `tile.pyrotech.masonry_brick_block.name`